Do homework4
