+++ scrapper/src/test/java/edu/java/scrapper/MigrationTest.java
@@ -1,0 +1,12 @@
+package edu.java.scrapper;
+
+import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+public class MigrationTest extends IntegrationEnvironment {
+
+    @Test
+    public void migrationRunningTest() {
+        assertTrue(IntegrationEnvironment.POSTGRES.isRunning());
+    }
+}
